:hammer: Fixed type index compilation error
https://github.com/Gjacquenot/test_clang/runs/7746125643?check_suite_focus=true

[613/908] Building CXX object grpc/CMakeFiles/grpc_airy.dir/src/AiryGRPC.cpp.o
FAILED: grpc/CMakeFiles/grpc_airy.dir/src/AiryGRPC.cpp.o
/usr/bin/clang++ -DGTEST_HAS_POSIX_RE=0 -I. -I/opt/share/code/grpc/inc -I/opt/share/code/ssc -I/opt/share/code/exceptions/inc -I/opt/share/code/external_data_structures/inc -I/opt/share/code/core/inc -I/opt/share/code/environment_models/inc -I/opt/share/code/hdb_interpolators/inc -I/opt/share/code/mesh/inc -isystem /opt/eigen -isystem /opt/boost/include -isystem /opt/grpc/include -isystem grpc -isystem /opt/share/code/yaml-cpp/src -isystem /opt/share/code/yaml-cpp/include -isystem /opt/share/code/yaml-cpp/include/yaml-cpp -Werror -pedantic -pedantic-errors -Wextra -Wall -Wunused-function -Wunused-label -Wunused-parameter -Wunused-value -Wunused-variable -fno-common -Wformat=2 -Winit-self -Wpacked -Wpointer-arith -Wmissing-declarations -Wmissing-format-attribute -Wsign-compare -Wstrict-aliasing=2 -Wundef -Wconversion -Wno-misleading-indentation -Woverloaded-virtual -Weffc++ -Wwrite-strings -Wfatal-errors -Wno-deprecated -Wvariadic-macros -fPIC -std=gnu++11 -O3 -DNDEBUG -MD -MT grpc/CMakeFiles/grpc_airy.dir/src/AiryGRPC.cpp.o -MF grpc/CMakeFiles/grpc_airy.dir/src/AiryGRPC.cpp.o.d -o grpc/CMakeFiles/grpc_airy.dir/src/AiryGRPC.cpp.o -c /opt/share/code/grpc/src/AiryGRPC.cpp
/opt/share/code/grpc/src/AiryGRPC.cpp:53:48: fatal error: implicit conversion changes signedness: 'int' to 'std::vector::size_type' (aka 'unsigned long') [-Wsign-conversion]
                reply->add_z(res_elevations.at(index));
                                            ~~ ^~~~~
1 error generated.

+++ code/grpc/src/AiryGRPC.cpp
@@ -50,7 +50,7 @@ class WavesImpl final : public Waves::Service {
             {
                 reply->add_x(request->x(index));
                 reply->add_y(request->y(index));
-                reply->add_z(res_elevations.at(index));
+                reply->add_z(res_elevations.at(static_cast<size_t>(index)));
             }
             return Status::OK;
         }
@@ -70,7 +70,7 @@ class WavesImpl final : public Waves::Service {
                 reply->add_x(request->x(index));
                 reply->add_y(request->y(index));
                 reply->add_z(request->z(index));
-                reply->add_pdyn(res_pdyn.at(index));
+                reply->add_pdyn(res_pdyn.at(static_cast<size_t>(index)));
             }
             return Status::OK;
         }
